improve allocation of large storage arrays for each supported type
git-svn-id: http://code.imagej.net/svn/imagej/trunk@1934 be23795f-a60e-40d4-ba19-055798872e57

+++ ij2-common/src/main/java/imagej/data/BitType.java
@@ -94,7 +94,7 @@ public class BitType implements Type
 	}
 
 	@Override
-	public Object allocateStorageArray(int numPixels)
+	public Object allocateStorageArray(long numPixels)
 	{
 		long numStorageUnits = calcNumStorageUnitsFromPixelCount(numPixels);
